+++ CHANGELOG.md
@@ -10,7 +10,8 @@ SPDX-License-Identifier: GPL-3.0-or-later
 
 ### Added
 
-- Add rules "spelling-ctxt", "spelling-id", "spelling-str" and "spelling"
+- Add rules "spelling-ctxt", "spelling-id" and "spelling-str"
+- Add special rule "spelling" to select all spelling rules
 - Add special rule "checks" to select all rules except fuzzy, obsolete and untranslated
 
 ### Fixed